add download function to datatable component and clean up styling
