+++ sites/example-project/src/pages/query-chain-test.md
@@ -10,11 +10,21 @@ select
 from `bigquery-public-data.austin_311.311_service_requests` 
 where created_date >= timestamp_sub(current_timestamp(), interval 180 day)
 group by 1,2 
+limit 5
 ```
+
+Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. 
+
+<DataTable data={data.input}/>
+
+Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. 
 
 ```working_reference 
     select count(*) as n_days from ${input}
 ```
+
+<DataTable data={data.working_reference}/>
+
 
 ```two_step_reference 
     select
